// src/pages/Services/NewService/interfaces/PersonCRMInterface.ts
import { IPersonAddressCRM } from "./PersonAddressCRMInterface"

export interface IPersonCRM {
  personId: number
  personGuid: string
  name: string
  shortName: string
  personType: number
  isActive: boolean
  documentCNPJ: string
  documentIE: string
  documentIM: string
  documentCPF: string
  documentRG: string
  isMale: boolean
  bornDate: string
  foundationDate: string
  defaultEmail: string
  phone: string
  cellPhone: string
  site: string
  defaultEmailGuid: string
  phoneGuid: string
  cellPhoneGuid: string
  siteGuid: string
  observation: string
  isFranchisee: boolean
  isSeller: boolean
  isCustomer: boolean
  isCollaborator: boolean
  franchiseeId: string
  availableCapital: string
  royaltyType: string
  meetFranchise: string
  regionId: string
  regionName: string
  regionDescription: string
  createDateTime: string
  regionStatus: string
  customerId: string
  collaboratorId: number
  sellerId: string
  userId: number
  username: string
  supervisorId: string
  supervisorName: string
  responsibleName: string
  responsibleFranchiseeId: string
  responsibleFranchiseeName: string
  ownerId: number
  isDeleted: boolean
  cityName: string
  stateAbbreviation: string
  stateName: string
  addressDescription: string
  addressName: string
  addressNumber: string
  canBeUpdated: boolean
  canBeDeleted: boolean
  parentOwnerId: number
  responsibleFranchiseeOwnerId: string
  personOwnerId: string
  guidPersonId: string
  qualificacaoId: string
  qualificacaoDescricao: string
  origemContatoId: string
  origemContatoDescricao: string
  obrigarOrganizacaoCEP: boolean
  doctoNaoPreenchido: number
  motivoDoctoNaoPreenchido: string
  personContacts: any[]
  personAddresses: IPersonAddressCRM[]
  segmentosMercado: any[]
  areasNegocio: any[]
}